feat: agregar endpoint para registrar colaboradores

+++ src/Domain/Entities/Payroll/Vacation.cs
@@ -5,9 +5,9 @@ namespace ERP.Core.Manager.Api.Domain.Entities.Payroll
     public class Vacation : BaseEntity<Guid>
     {
         required public Guid CollaboratorId { get; set; }
-        public float AvailableVacations { get; set; }
-        public float GeneredVacation { get; set; }
-        public float EnjoyedVacation { get; set; }
+        public decimal AvailableVacations { get; set; }
+        public decimal GeneredVacation { get; set; }
+        public decimal EnjoyedVacation { get; set; }
         
         public virtual Collaborator Collaborator { get; set; } = null!;
     }
